Crate core_crypto_keystore

Re-exports

pub use self::connection::ConnectionType;

pub use self::connection::Database;

Modules

connection

entities

traits

transaction

unified_connection

Structs

DatabaseKey

The key used to encrypt the database.

Sha256Hash

A Sha256 hash.

Enums

CryptoKeystoreError

Error type to represent various errors that can happen in the KeyStore

Traits

CryptoKeystoreMls

An interface for the specialized queries in the KeyStore

CryptoKeystoreProteus

Functions

deser

ser

Type Aliases

CryptoKeystoreResult

A specialized Result for the KeyStore functions