core_crypto_keystore

Module ancillary

Types and functions which support the entities (i.e. as primary or search keys) without themselves being entities.

Structs

ConversationEpochsOlderThan

Some data, e.g., TntSecrets, need to be cleaned up if they belong to a conversation epoch that is older than a given epoch. To enable this, SearchableEntity is implemented with this search key.

ConversationId

Type-safe reference to a conversation id.

ConversationIdRef

Reference to a ConversationId.

MessageRxCounterPk

Owned key shared by the targeted and transient message receive counters.

MessageRxCounterPkRef

Borrowed form of MessageRxCounterPk.