Module core_crypto_keystore::connection

Re-exports

pub use self::platform::*;

Modules

platform

Structs

Connection

Constants

MAX_BLOB_LEN

Limit on the length of a blob to be stored in the database. This limit applies to both SQLCipher-backed stores and WASM. This limit is conservative on purpose when targeting WASM, as the lower bound that exists is Safari with a limit of 1GB per origin.

Traits

DatabaseConnection