Module core_crypto_keystore::entities

Modules

credential

e2ei_acme_ca

e2ei_crl

e2ei_intermediate_cert

encryption_keypair

enrollment

epoch_encryption_keypair

group

hpke_private_key

keypackage

pending_group

pending_message

prekey

psk_bundle

refresh_token

session

signature_keypair

Structs

E2eiAcmeCA

E2eiCrl

E2eiEnrollment

Entity representing an enrollment instance used to fetch a x509 certificate and persisted when context switches and the memory it lives in is about to be erased

E2eiIntermediateCert

E2eiRefreshToken

OIDC refresh token used in E2EI

EntityFindParams

MlsCredential

Entity representing a persisted Credential

MlsEncryptionKeyPair

Entity representing a persisted HpkePrivateKey (related to LeafNode Private keys that the client is aware of)

MlsEpochEncryptionKeyPair

Entity representing a list of MlsEncryptionKeyPair

MlsHpkePrivateKey

Entity representing a persisted HpkePrivateKey (related to LeafNode Private keys that the client is aware of)

MlsKeyPackage

Entity representing a persisted KeyPackage

MlsPendingMessage

Entity representing a buffered message

MlsPskBundle

Entity representing a persisted SignatureKeyPair

MlsSignatureKeyPair

Entity representing a persisted SignatureKeyPair

PersistedMlsGroup

Entity representing a persisted MlsGroup

PersistedMlsPendingGroup

Entity representing a temporarily persisted MlsGroup

ProteusIdentity

ProteusPrekey

ProteusSession

StringEntityId

Traits

Entity

EntityBase

EntityIdStringExt

MlsCredentialExt

PersistedMlsGroupExt

UniqueEntity