core_crypto_keystore::entities

Struct ConsumerData

pub struct ConsumerData {
    pub content: Vec<u8>,
}

Consumers of this library can use this to specify data to be persisted at the end of a transaction.

Fields
content: Vec<u8>