core_crypto_keystore::entities

Trait Entity

pub trait Entity: EntityBase {
    // Required methods
    fn id_raw(&self) -> &[u8] ;
    fn find_all<'life0, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        params: EntityFindParams,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn find_one<'life0, 'life1, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        id: &'life1 StringEntityId<'_>,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Option<Self>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn count<'life0, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn merge_key(&self) -> Vec<u8>  { ... }
    fn find_many<'life0, 'life1, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        ids: &'life1 [StringEntityId<'_>],
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Required Methods

fn id_raw(&self) -> &[u8]

fn find_all<'life0, 'async_trait>( conn: &'life0 mut Self::ConnectionType, params: EntityFindParams, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn find_one<'life0, 'life1, 'async_trait>( conn: &'life0 mut Self::ConnectionType, id: &'life1 StringEntityId<'_>, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Option<Self>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn count<'life0, 'async_trait>( conn: &'life0 mut Self::ConnectionType, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn merge_key(&self) -> Vec<u8>

The query results that are obtained during a transaction from the transaction cache and the database are merged by this key.

fn find_many<'life0, 'life1, 'async_trait>( conn: &'life0 mut Self::ConnectionType, ids: &'life1 [StringEntityId<'_>], ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.