core_crypto_keystore

Struct UniqueArc

pub struct UniqueArc<T> { /* private fields */ }

A smart pointer which has exactly one strong reference to the inner type, but may have several weak references.

This type is intentionally !Clone so that we know for a fact that it can be safely unpacked.

Implementations

impl UniqueArc<Transaction>

pub async fn commit(self) -> Result<(), CryptoKeystoreError>

Persists all the operations in the database. It will effectively open a transaction internally, perform all the buffered operations and commit.

impl<T> UniqueArc<T>

pub async fn into_inner(this: Self) -> T

Unpack this UniqueArc, returning the inner value.

This waits for any in-flight UniqueWeak::upgrade / UniqueWeak::upgrade_sync guard to drop before unpacking, so that the sole remaining strong reference is the one held by this UniqueArc.

pub fn downgrade(this: &Self) -> UniqueWeak<T>

Produce a weak reference to this item