Trait core_crypto_keystore::CryptoKeystoreMls

pub trait CryptoKeystoreMls: Sized {
    fn mls_load_identity_signature<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Option<Vec<u8>>>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn mls_save_identity_signature<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        id: &'life1 str,
        signature: &'life2 [u8],
        credential: &'life3 [u8]
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;
    fn mls_keypackagebundle_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<usize>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn mls_fetch_keypackage_bundles<'life0, 'async_trait, V>(
        &'life0 self,
        count: u32
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<V>>> + 'async_trait>>
    where
        V: 'async_trait + FromKeyStoreValue,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn mls_get_keypackage<'life0, 'async_trait, V>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<V>> + 'async_trait>>
    where
        V: 'async_trait + FromKeyStoreValue,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn mls_group_persist<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        group_id: &'life1 [u8],
        state: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn mls_groups_restore<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<HashMap<Vec<u8>, Vec<u8>>>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn mls_group_delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        group_id: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn mls_pending_groups_save<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        group_id: &'life1 [u8],
        mls_group: &'life2 [u8],
        custom_configuration: &'life3 [u8]
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        Self: 'async_trait;
    fn mls_pending_groups_load<'life0, 'life1, 'async_trait>(
        &'life0 self,
        group_id: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<(Vec<u8>, Vec<u8>)>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn mls_pending_groups_delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        group_id: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

An interface for the specialized queries in the KeyStore

Required Methods

fn mls_load_identity_signature<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 str
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Option<Vec<u8>>>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Retrieves an identity signature from a client id

Arguments

id - client id

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_save_identity_signature<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
id: &'life1 str,
signature: &'life2 [u8],
credential: &'life3 [u8]
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Saves the signature and credentials for a client id

Arguments

id - client id

signature - the signature to be stored

credential - the credential to be stored

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_keypackagebundle_count<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<usize>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Counts how many KeyPackages are stored

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_fetch_keypackage_bundles<'life0, 'async_trait, V>(
&'life0 self,
count: u32
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<V>>> + 'async_trait>>where
V: 'async_trait + FromKeyStoreValue,
'life0: 'async_trait,
Self: 'async_trait,

Fetches Keypackages

Arguments

count - amount of entries to be returned

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_get_keypackage<'life0, 'async_trait, V>(
&'life0 self
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<V>> + 'async_trait>>where
V: 'async_trait + FromKeyStoreValue,
'life0: 'async_trait,
Self: 'async_trait,

Fetches a singles keypackage

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_group_persist<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
group_id: &'life1 [u8],
state: &'life2 [u8]
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Persists a MlsGroup

Arguments

group_id - group/conversation id

state - the group state

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_groups_restore<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<HashMap<Vec<u8>, Vec<u8>>>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Loads MlsGroups from the database. It will be returned as a HashMap where the key is the group/conversation id and the value the group state

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_group_delete<'life0, 'life1, 'async_trait>(
&'life0 self,
group_id: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Deletes MlsGroups from the database.

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_pending_groups_save<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
group_id: &'life1 [u8],
mls_group: &'life2 [u8],
custom_configuration: &'life3 [u8]
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
Self: 'async_trait,

Saves a MlsGroup in a temporary table (typically used in scenarios where the group cannot be committed until the backend acknowledges it, like external commits)

Arguments

group_id - group/conversation id

mls_group - the group/conversation state

custom_configuration - local group configuration

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_pending_groups_load<'life0, 'life1, 'async_trait>(
&'life0 self,
group_id: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<(Vec<u8>, Vec<u8>)>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Loads a temporary MlsGroup and its configuration from the database

Arguments

id - group/conversation id

Errors

Any common error that can happen during a database connection. IoError being a common error for example.

fn mls_pending_groups_delete<'life0, 'life1, 'async_trait>(
&'life0 self,
group_id: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Deletes a temporary MlsGroup from the database

Arguments

id - group/conversation id

Errors

Any common error that can happen during a database connection. IoError being a common error for example.