core_crypto_keystore

Module traits

Traits

DecryptData

This trait is intended to provide a convenient way to decrypt data.

Decryptable

Helper trait for restoring from an encrypted form of this struct.

Decrypting

This trait restores to a plaintext form of this struct, where all sensitive fields have been decrypted.

EncryptData

This trait is intended to provide a convenient way to encrypt data.

Encrypting

This trait produces an encrypted form of this struct, where all sensitive fields have been encrypted using the EncryptData helper.

Entity

Something which can be stored in our database.

EntityBase

A supertrait that all entities must implement. This handles multiplexing over the two different database backends.

EntityGetBorrowed

An extension trait which should be implemented for all entities whose primary key has a distinct borrowed form.

EntityTransactionDeleteBorrowed

Extend an Entity with db-mutating operations which can be performed when provided with a transaction.

EntityTransactionExt

Extend an Entity with db-mutating operations which can be performed when provided with a transaction.

FetchFromDatabase

Interface to fetch from the database either from the connection directly or through a transaction.

KeyType

A Key Type is a type which can act as a key for a database.

UniqueEntity

A unique entity can appear either 0 or 1 times in the database.

UniqueEntityExt

Unique entities get some convenience methods implemented automatically.