core_crypto_keystore::traits

Trait Decrypting

pub trait Decrypting<'de>: 'de + Deserialize<'de> {
    type DecryptedForm: Entity;

    // Required method
    fn decrypt(
        self,
        cipher: &Aes256Gcm,
    ) -> CryptoKeystoreResult<Self::DecryptedForm>;
}

This trait restores to a plaintext form of this struct, where all sensitive fields have been decrypted.

This is quite likely to be handled automatically by a macro, depending on how annoying it is to implement everywhere.

Example

// Foo is an Entity
struct Foo {
    id: Vec<u8>,
    sensitive_data: Vec<u8>, // sensitive!
}

#[derive(serde::Serialize)]
struct EncryptedFoo<'de> {
    id: Vec<u8>,
    sensitive_data: &'de [u8],
}

impl<'de> Decrypting<'de> for EncryptedFoo<'de> {
    type DecryptedForm = Foo;

    fn decrypt(self, cipher: &aes_gcm::Aes256Gcm) -> CryptoKeystoreResult<Foo> {
        let id = self.id;
        let sensitive_data = E::decrypt_data(cipher, &id, self.sensitive_data)?;
        Ok(Foo {
            id,
            sensitive_data,
        })
    }
}

This can then be used like:

let foo = serde_json::from_str::<EncryptedFoo>(json)?.decrypt(cipher)?;

Required Associated Types

type DecryptedForm: Entity

Required Methods

fn decrypt( self, cipher: &Aes256Gcm, ) -> CryptoKeystoreResult<Self::DecryptedForm>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.