core_crypto_keystore::traits

Trait EncryptionKey

pub trait EncryptionKey {
    // Required method
    fn encryption_key(&self) -> &[u8] ;
}

This trait is a hack enabling us to encrypt types for which we don’t use the primary key in the AAD.

The only reason we’d ever want this is if the primary key is not what we actually use, and the only reason that would be the case is if we’re faking a primary key where no such key really exists.

In other words, MLS pending messages.

Required Methods

fn encryption_key(&self) -> &[u8]

Get the key bytes which are to be used as the encryption key for this data.