core_crypto_keystore::traits

Trait EntityBase

pub trait EntityBase: 'static + Sized {
    type ConnectionType: for<'a> DatabaseConnection<'a>;
    type AutoGeneratedFields: Default;

    const COLLECTION_NAME: &'static str;

    // Required method
    fn to_transaction_entity(self) -> Entity;

    // Provided method
    fn downcast<T: EntityBase>(&self) -> Option<&T> { ... }
}

A supertrait that all entities must implement. This handles multiplexing over the two different database backends.

This trait should be removed once the persistence layers are unified. See WPB-16241.

Required Associated Constants

const COLLECTION_NAME: &'static str

Beware: if you change the value of this constant on any WASM entity, you’ll need to do a data migration not only because it is used as reference to the object store names but also for the value of the aad.

Required Associated Types

type ConnectionType: for<'a> DatabaseConnection<'a>

type AutoGeneratedFields: Default

Required Methods

fn to_transaction_entity(self) -> Entity

Provided Methods

fn downcast<T: EntityBase>(&self) -> Option<&T>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.