core_crypto_keystore::traits

Trait KeyType

pub trait KeyType: Send + Sync {
    // Required method
    fn bytes(&self) -> Cow<'_, [u8]>;
}

A Key Type is a type which can act as a key for a database.

This might be a primary key, in which case the key uniquely identifies either 0 or 1 entries in the database. Or it might be a search key, in which case the key could match any number of entries.

Required Methods

fn bytes(&self) -> Cow<'_, [u8]>

Get a unique binary representation of this key.

For simple keys it can just be the borrowed form of the key itself, but for complex keys it could be the run length encoding of each field of the key.