core_crypto_keystore::traits

Trait OwnedKeyType

pub trait OwnedKeyType: 'static + KeyType {
    // Required method
    fn from_bytes(bytes: &[u8]) -> Option<Self>;
}

An owned key type can be converted to from arbitrary bytes.

Required Methods

fn from_bytes(bytes: &[u8]) -> Option<Self>

Parse some bytes into an instance of this type.

We’re just going with Option instead of CryptoKeystoreResult for now because the hopeful assumption is that this is going to be a rare occurrence that doesn’t need much explanation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.