core_crypto_keystore::traits

Trait UnifiedSearchableEntity

pub trait UnifiedSearchableEntity<SearchKey: KeyType>: UnifiedEntity {
    // Required methods
    fn find_all_matching(
        conn: &Connection,
        search_key: &SearchKey,
    ) -> CryptoKeystoreResult<Vec<Self>>;
    fn matches(&self, search_key: &SearchKey) -> bool;
}

Required Methods

fn find_all_matching( conn: &Connection, search_key: &SearchKey, ) -> CryptoKeystoreResult<Vec<Self>>

Find all entities matching the search key.

The specific meaning of “matching” the search key will depend on the entity in question, but generally the search key will have one or more fields which effectively act as a WHERE-clause for the search.

fn matches(&self, search_key: &SearchKey) -> bool

true when this entity instance matches the search key.

The specific meaning of “matching” the search key will depend on the entity in question, but generally the search key will have one or more fields which must equal some fields on the entity.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.