core_crypto_keystore

Module connection

Re-exports

pub use self::platform::*;

Modules

platform

Structs

ConnectionGuard

Exclusive access to the database connection

Database

DatabaseKey

The key used to encrypt the database.

Enums

ConnectionType

Where to open a connection

Constants

MAX_BLOB_LEN

Limit on the length of a blob to be stored in the database.

Traits

DatabaseConnection

DatabaseConnectionRequirements

FetchFromDatabase

Interface to fetch from the database either from the connection directly or through a transaaction