core_crypto_keystore

Module entities

Modules

credential

e2ei_acme_ca

encryption_keypair

group

hpke_private_key

pending_group

pending_message

prekey

psk_bundle

Structs

ConsumerData

Consumers of this library can use this to specify data to be persisted at the end of a transaction.

E2eiAcmeCA

E2eiCrl

E2eiCrlDecrypt

E2eiCrlEncrypt

E2eiIntermediateCert

E2eiIntermediateCertDecrypt

E2eiIntermediateCertEncrypt

E2eiRefreshToken

OIDC refresh token used in E2EI

EntityFindParams

MlsPendingMessage

Entity representing a buffered message

MlsPendingMessagePrimaryKey

MlsPendingMessages have no distinct primary key; they must always be accessed via MlsPendingMessage::find_all_by_conversation_id and cleaned up with MlsPendingMessage::delete_by_conversation_id

PersistedMlsGroup

Entity representing a persisted MlsGroup

PersistedMlsGroupDecrypt

PersistedMlsGroupEncrypt

PersistedMlsPendingGroup

Entity representing a temporarily persisted MlsGroup

ProteusIdentity

ProteusPrekey

ProteusSession

ProteusSessionDecrypt

ProteusSessionEncrypt

StoredBufferedCommit

Entity representing a buffered commit.

StoredBufferedCommitDecrypt

StoredBufferedCommitEncrypt

StoredCredential

Entity representing a persisted Credential

StoredE2eiEnrollment

Entity representing an enrollment instance used to fetch a x509 certificate and persisted when context switches and the memory it lives in is about to be erased

StoredE2eiEnrollmentDecrypt

StoredE2eiEnrollmentEncrypt

StoredEncryptionKeyPair

Entity representing a persisted HpkePrivateKey (related to LeafNode Private keys that the client is aware of)

StoredEpochEncryptionKeypair

Entity representing a list of StoredEncryptionKeyPair

StoredEpochEncryptionKeypairDecrypt

StoredEpochEncryptionKeypairEncrypt

StoredHpkePrivateKey

Entity representing a persisted HpkePrivateKey (related to LeafNode Private keys that the client is aware of)

StoredKeypackage

Entity representing a persisted KeyPackage

StoredKeypackageDecrypt

StoredKeypackageEncrypt

StoredPskBundle

Entity representing a persisted SignatureKeyPair

StringEntityId

Traits

Entity

EntityBase

EntityIdStringExt

EntityTransactionExt

PersistedMlsGroupExt

UniqueEntity