Trait core_crypto_keystore::entities::EntityBase

pub trait EntityBase:
    Send
    + Sized
    + Clone
    + PartialEq
    + Eq
    + Debug {
    type ConnectionType: DatabaseConnection;
    type AutoGeneratedFields;

    const COLLECTION_NAME: &'static str;

    // Required methods
    fn to_missing_key_err_kind() -> MissingKeyErrorKind;
    fn save<'life0, 'life1, 'async_trait>(
        &'life0 self,
        conn: &'life1 mut Self::ConnectionType,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn find_all<'life0, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        params: EntityFindParams,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn find_one<'life0, 'life1, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        id: &'life1 StringEntityId<'_>,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Option<Self>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn count<'life0, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete<'life0, 'life1, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        id: &'life1 [StringEntityId<'_>],
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn insert<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _conn: &'life1 mut Self::ConnectionType,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Self::AutoGeneratedFields>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn find_many<'life0, 'life1, 'async_trait>(
        conn: &'life0 mut Self::ConnectionType,
        ids: &'life1 [StringEntityId<'_>],
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Required Associated Types

type ConnectionType: DatabaseConnection

type AutoGeneratedFields

Required Associated Constants

const COLLECTION_NAME: &'static str

Beware: if you change the value of this constant on any WASM entity, you’ll need to do a data migration not only because it is used as reference to the object store names but also for the value of the aad.

Required Methods

fn to_missing_key_err_kind() -> MissingKeyErrorKind

fn save<'life0, 'life1, 'async_trait>( &'life0 self, conn: &'life1 mut Self::ConnectionType, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn find_all<'life0, 'async_trait>( conn: &'life0 mut Self::ConnectionType, params: EntityFindParams, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn find_one<'life0, 'life1, 'async_trait>( conn: &'life0 mut Self::ConnectionType, id: &'life1 StringEntityId<'_>, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Option<Self>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn count<'life0, 'async_trait>( conn: &'life0 mut Self::ConnectionType, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn delete<'life0, 'life1, 'async_trait>( conn: &'life0 mut Self::ConnectionType, id: &'life1 [StringEntityId<'_>], ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Provided Methods

fn insert<'life0, 'life1, 'async_trait>( &'life0 self, _conn: &'life1 mut Self::ConnectionType, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Self::AutoGeneratedFields>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn find_many<'life0, 'life1, 'async_trait>( conn: &'life0 mut Self::ConnectionType, ids: &'life1 [StringEntityId<'_>], ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<Vec<Self>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Object Safety

This trait is not object safe.