core_crypto_keystore::traits

Trait EntityDeleteBorrowed

pub trait EntityDeleteBorrowed: EntityDatabaseMutation + BorrowPrimaryKey {
    // Required method
    fn delete_borrowed(
        tx: &Transaction<'_>,
        id: Self::BorrowedPrimaryKey<'_>,
    ) -> CryptoKeystoreResult<bool>;
}

Extend an Entity with db-mutating operations which can be performed when provided with a transaction.

Required Methods

fn delete_borrowed( tx: &Transaction<'_>, id: Self::BorrowedPrimaryKey<'_>, ) -> CryptoKeystoreResult<bool>

Delete an entity by a borrowed form of its primary key.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".