core_crypto_keystore::traits

Trait EntityDeleteBorrowed

pub trait EntityDeleteBorrowed<'a>: EntityDatabaseMutation<'a> + BorrowPrimaryKey {
    // Required method
    fn delete_borrowed<'life0, 'life1, 'async_trait>(
        tx: &'life0 Self::Transaction,
        id: &'life1 Self::BorrowedPrimaryKey,
    ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<bool>> + Send + 'async_trait>>
       where for<'pk> &'pk Self::BorrowedPrimaryKey: KeyType,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Extend an Entity with db-mutating operations which can be performed when provided with a transaction.

Required Methods

fn delete_borrowed<'life0, 'life1, 'async_trait>( tx: &'life0 Self::Transaction, id: &'life1 Self::BorrowedPrimaryKey, ) -> Pin<Box<dyn Future<Output = CryptoKeystoreResult<bool>> + Send + 'async_trait>>
where for<'pk> &'pk Self::BorrowedPrimaryKey: KeyType, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Delete an entity by a borrowed form of its primary key.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.