core_crypto_keystore::traits

Trait PrimaryKey

pub trait PrimaryKey {
    type PrimaryKey: OwnedKeyType;

    // Required method
    fn primary_key(&self) -> Self::PrimaryKey;
}

Something which has a distinct primary key which can uniquely identify it.

Required Associated Types

type PrimaryKey: OwnedKeyType

Each distinct PrimaryKey uniquely identifies either 0 or 1 instance.

This constraint should be enforced at the DB level.

Required Methods

fn primary_key(&self) -> Self::PrimaryKey

Get this entity’s primary key.

This must return an owned type, because there are some entities for which only owned primary keys are possible. However, entities which have primary keys owned within the entity itself should consider also implementing BorrowPrimaryKey for greater efficiency.