core_crypto_keystore::transaction

Struct Transaction

pub struct Transaction { /* private fields */ }

This is a guard over an in-flight transaction.

In a perfect world we’d be able to use [rusqlite::Transaction], but that type is intentionally !Send + !Sync, which prevents us from being able to keep a long-lived transaction around.

Dropping the transaction without committing performs an implicit rollback.

This type is always wrapped in a UniqueArc, which keeps things efficient, at the cost of prohibiting Clone. In case you need to share this around, there are weak references available via UniqueArc::downgrade. Alternately, wrap the entire thing in an Arc<Mutex<Option<UniqueArc<Self>>>> or similar. Just be aware that you’ll need to take the unique arc out in order to commit.

Implementations

impl Transaction

pub fn conn(&self) -> CryptoKeystoreResult<TransactionConnection>

Guard over this transaction’s connection.

Errors if SQLite has already rolled this transaction back under us.

impl Transaction

pub async fn with_savepoint<T, E>( &self, savepoint_name: &str, operation: impl AsyncFnOnce() -> Result<T, E>, map_err: impl Fn(&'static str) -> Box<dyn FnOnce(CryptoKeystoreError) -> E>, ) -> Result<T, E>

Do an operation within the context of a savepoint.

https://sqlite.org/lang_savepoint.html

A savepoint is like a transaction, but has a name and can be nested. This allows for more fine-grained control over what makes it into the history, or not.

The operation produces an arbitrary result which is forwarded unmodified to the output of this method. If that result is ok, the savepoint is released (i.e. committed). If that result is an error, the savepoint is rolled back.

Error-mapping is the most complicated part of this. When an error is produced in the course of this method, we run the map_err function with a static context string describing what was happening. That function must return a second function which transforms a CryptoKeystoreError into an instance of E.